+++ src/components/form/Presentation/index.tsx
@@ -1,4 +1,3 @@
-import CancelIcon from '@assets/svg/x.svg';
 import ApplicationPeriodField from '@components/form/Presentation/ApplicationPeriodField';
 import CategoryField from '@components/form/Presentation/CategoryField';
 import CoLeaderField from '@components/form/Presentation/CoLeaderField';
@@ -7,14 +6,13 @@ import useImageHandler from '@components/form/Presentation/ImageField/useImageHa
 import KeywordField from '@components/form/Presentation/KeywordField';
 import LeaderDescriptionField from '@components/form/Presentation/LeaderDescriptionField';
 import NoticeField from '@components/form/Presentation/NoticeField';
+import SubmitPresentationButton from '@components/form/Presentation/SubmitPresentationButton';
 import TargetField from '@components/form/Presentation/TargetField';
 import TitleField from '@components/form/Presentation/TitleField';
 import WelcomeMessageField from '@components/form/Presentation/WelcomeMessageField';
 import { colors } from '@sopt-makers/colors';
 import { fontsObject } from '@sopt-makers/fonts';
-import { useDialog } from '@sopt-makers/ui';
-import { useRouter } from 'next/router';
-import React, { ReactNode, useRef } from 'react';
+import React, { useRef } from 'react';
 import { styled } from 'stitches.config';
 import ActivityPeriodField from './ActivityPeriodField';
 import DescriptionField from './DescriptionField';
@@ -27,17 +25,6 @@ interface PresentationProps {
   onSubmit: React.FormEventHandler<HTMLFormElement>;
   disabled?: boolean;
 }
-interface TypeOptionsProp {
-  cancelButtonText?: string;
-  approveButtonText?: string;
-  buttonFunction?: () => void;
-}
-interface DialogOptionType {
-  title: ReactNode;
-  description: ReactNode;
-  type?: 'default' | 'danger' | 'single' | undefined;
-  typeOptions?: TypeOptionsProp;
-}
 
 function Presentation({
   submitButtonLabel,
@@ -47,31 +34,12 @@ function Presentation({
   onSubmit,
   disabled = true,
 }: PresentationProps) {
-  const router = useRouter();
-  const { open } = useDialog();
   const { handleChangeFile, handleDeleteFile, handleAddFiles } = useImageHandler({
     onChangeImage: handleChangeImage,
     onDeleteImage: handleDeleteImage,
   });
 
-  const isEdit = router.asPath.includes('/edit');
-
   const formRef = useRef<HTMLFormElement>(null);
-
-  const dialogOption: DialogOptionType = {
-    title: `모임을 ${isEdit ? '수정' : '개설'}하시겠습니까?`,
-    description: '모임에 대한 설명이 충분히 작성되었는지 확인해 주세요',
-    type: 'default',
-    typeOptions: {
-      cancelButtonText: '취소',
-      approveButtonText: `${isEdit ? '수정' : '개설'}하기`,
-      buttonFunction: () => {
-        if (formRef.current) {
-          formRef.current.requestSubmit();
-        }
-      },
-    },
-  };
 
   return (
     <SForm onSubmit={onSubmit} ref={formRef}>
@@ -110,25 +78,12 @@ function Presentation({
           <NoticeField />
         </div>
       </div>
-
-      {/* TODO: icon이 포함된 컴포넌트를 주입받아야 한다. */}
-      <ButtonContainer>
-        {cancelButtonLabel && (
-          <Button type="button" onClick={() => router.back()}>
-            <CancelIcon />
-            {cancelButtonLabel}
-          </Button>
-        )}
-        <SubmitButton
-          type="button"
-          onClick={() => {
-            open(dialogOption);
-          }}
-          disabled={disabled}
-        >
-          {submitButtonLabel}
-        </SubmitButton>
-      </ButtonContainer>
+      <SubmitPresentationButton
+        cancelButtonLabel={cancelButtonLabel}
+        submitButtonLabel={submitButtonLabel}
+        disabled={disabled}
+        formRef={formRef}
+      />
     </SForm>
   );
 }
@@ -141,44 +96,6 @@ const SForm = styled('form', {
   gap: '60px',
   '@media (max-width: 768px)': {
     gap: '56px',
-  },
-});
-
-const ButtonContainer = styled('div', {
-  display: 'flex',
-  gap: '20px',
-  alignSelf: 'flex-end',
-
-  '@media (max-width: 768px)': {
-    flexDirection: 'column-reverse',
-    width: '100%',
-    marginBottom: '20px',
-    gap: '16px',
-  },
-});
-const Button = styled('button', {
-  padding: '16px 20px',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  gap: '12px',
-  background: '$gray600',
-  borderRadius: '10px',
-  fontAg: '18_bold_100',
-  color: '$gray10',
-
-  '@media (max-width: 768px)': {
-    gap: '10px',
-    width: '100%',
-    fontAg: '16_bold_100',
-  },
-});
-const SubmitButton = styled(Button, {
-  background: '$gray10',
-  color: '$gray950',
-  '&:disabled': {
-    cursor: 'not-allowed',
-    opacity: 0.35,
   },
 });
